+++ htslib/htslib/sam.h
@@ -54,7 +54,7 @@ extern "C" {
 typedef struct {
     int32_t n_targets, ignore_sam_err;
     uint32_t l_text;
-    uint32_t *target_len;
+    uint64_t *target_len;
     int8_t *cigar_tab;
     char **target_name;
     char *text;
